+++ src/components/pages/Changelog.tsx
@@ -3,7 +3,7 @@ import './Changelog.css';
 import _ from 'lodash';
 import React, { useEffect } from 'react';
 
-import * as Changes from '../Changes.md';
+import * as Changes from '../../Changes.md';
 
 function Changelog() { 
   useEffect(() => {